Log name & args of method that calls 'handle_api_response()'
Since this this mehtod is a helper method that is called by almost every other method in the file to parse the response from calling the Lightrail API, it is much more useful to log the name and arguments of the method from which it was called than to say that it triggered an error itself.

The new helper method 'getCallerAndCallerArgs()' uses PHP's built-in backtracer to to get the context 'one level up' from the context it is called in ('trace[2]') and extracts the function name and arguments, defaulting to 'global' if there was no calling function and an empty array if there were no arguments. These are returned as a PHP array for easy access.

This commit also updates the default error message that wraps the function name and arguments in 'get_printable_error_info()' to make it clearer that the message might be about a function that did not directly trigger an error.