Update READMEs: installation options; requirements
Add second option in installation instructions, as many WP users may not have access to their host's file system or be comfortable with it if they do.

Also mention the required versions of Wordpress, WooCommerce, and PHP7 for more immediate clarity (many WP users may still be running PHP5 so this is an important compatibility note).