Add dev and prod config files
Since configuration needs to be different for plugin development vs release, and it is likely that as the plugin evolves there will be more configuration than less, it makes sense to track and save these configs in separate files. These files have been created outside of the main plugin directory that will be packaged up and installed.
Currently the only difference between them is that the constants that point to Lightrail URIs (the API and the card details page in the webapp) should point to the production platform for the production version of the plugin, and to the dev platform for plugin development.