Barebones release script
Currently the release script's most important task is to update the plugin config such that it connects to the 'real' API & webapp instead of preprod, and does not produce random exceptions meant for testing. It does this by overwriting the plugin config with the contents of the 'prod config' file that lives outside the main plugin file.
It does not provide for switching back to the dev config: this script is meant to be run on a release branch only.