Change text occurrences of plugin name to 'Lightrail for WooCommerce'
Wordpress requires that plugins not use the name of another plugin as the first word of their name unless they are officially representing that plugin/brand/etc, so this name change helps us to comply with WP's guidelines.

Note this change only affects the text occurrences of the name: use of 'woocommerce-lightrail' and the like in filenames, directory names, and code has not been modified.

Also fixed a minor typo in the readme.md (changed 'paster' to 'paste' ln 51).