Move plugin files to root directory
This is the final step that will allow the entire project to be zipped/published as an installable plugin. Previously, the plugin files were in a subdirectory that needed to be zipped separately. Maintaining this file structure would have caused synchronisation headaches for releases, especially with regard to the WordPress SVN repository which requires plugins to adhere to a specific file structure.

Release script updated to reflect new location of plugin config file.