Fix subheaders in installation instructions
The subsection 'Upload and activate the plugin' was mistakenly formatted as a full section header, so it was parsed separately and not included in the installation instructions. This commit fixes that, and reformats the subheaders within that section to make it clear that they are part of the same section.