Add instructions for auto installation
Since our plugin is now live in the Wordpress.org plugin directory, the simplest installation option is to let Wordpress handle it.

Adds a third subsection with these installation instructions and modifies the subsection 'titles' for the other two to make it clear that you only need to do one before moving on.